requirements.txt
Flask==2.1.1
pandas==1.3.5
gevent==21.12.0
jinja2==3.0.3
python-dotenv==0.19.2
gunicorn==20.1.0
